Removes encryption from config file handling
This change removes the encryption functionality from the config file handling.

The previous implementation used encryption, which added complexity and potential overhead. The refactor simplifies the configuration management by directly using a local database without encryption.
